Use fixed size in place vector for object header messages
The max size for this is questionable, since it can get pretty large, especially with messages with many object header continuations. Currently, the max is 48.

To mitigate both this problem and the issue of manual stack size when traversing object header continuation messages, maybe the newer allocation sizes for messages should increase proportionally to how many messages already increase to reduce the depth?